+++ src/main/java/org/cotato/csquiz/api/attendance/controller/AttendanceController.java
@@ -13,7 +13,8 @@ import org.cotato.csquiz.api.attendance.dto.AttendancesResponse;
 import org.cotato.csquiz.api.attendance.dto.GenerationMemberAttendanceRecordResponse;
 import org.cotato.csquiz.api.attendance.dto.UpdateAttendanceRecordRequest;
 import org.cotato.csquiz.api.attendance.dto.UpdateAttendanceRequest;
-import org.cotato.csquiz.domain.attendance.service.AttendanceAdminService;
+import org.cotato.csquiz.domain.attendance.service.AttendanceExcelService;
+import org.cotato.csquiz.domain.attendance.service.AttendanceRecordService;
 import org.cotato.csquiz.domain.attendance.service.AttendanceService;
 import org.cotato.csquiz.domain.attendance.util.AttendanceExcelHeaderUtil;
 import org.springframework.http.HttpHeaders;
@@ -36,7 +37,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/v2/api/attendances")
 public class AttendanceController {
 
-    private final AttendanceAdminService attendanceAdminService;
+    private final AttendanceExcelService attendanceExcelService;
+    private final AttendanceRecordService attendanceRecordService;
     private final AttendanceService attendanceService;
 
     @Operation(summary = "출석 단건 조회")
@@ -48,7 +50,7 @@ public class AttendanceController {
     @Operation(summary = "출석 정보 변경 API")
     @PatchMapping
     public ResponseEntity<Void> updateAttendance(@RequestBody @Valid UpdateAttendanceRequest request) {
-        attendanceAdminService.updateAttendanceByAttendanceId(request);
+        attendanceService.updateAttendance(request.attendanceId(), request.location(), request.attendTime().attendanceDeadLine(), request.attendTime().lateDeadLine());
         return ResponseEntity.noContent().build();
     }
 
@@ -63,14 +65,14 @@ public class AttendanceController {
     public ResponseEntity<List<GenerationMemberAttendanceRecordResponse>> findAttendanceRecords(
             @RequestParam(name = "generationId") Long generationId
     ) {
-        return ResponseEntity.ok().body(attendanceAdminService.findAttendanceRecords(generationId));
+        return ResponseEntity.ok().body(attendanceRecordService.findAttendanceRecords(generationId));
     }
 
     @Operation(summary = "회원 출결사항 출석 단위 조회 API")
     @GetMapping("/{attendance-id}/records")
     public ResponseEntity<List<AttendanceRecordResponse>> findAttendanceRecordsByAttendance(
             @PathVariable("attendance-id") Long attendanceId) {
-        return ResponseEntity.ok().body(attendanceAdminService.findAttendanceRecordsByAttendance(attendanceId));
+        return ResponseEntity.ok().body(attendanceRecordService.findAttendanceRecordsByAttendance(attendanceId));
     }
 
     @Operation(summary = "회원 출결사항 수정 API")
@@ -78,7 +80,7 @@ public class AttendanceController {
     public ResponseEntity<Void> updateAttendanceRecords(
             @PathVariable("attendance-id") Long attendanceId,
             @RequestBody @Valid UpdateAttendanceRecordRequest request) {
-        attendanceAdminService.updateAttendanceRecords(attendanceId, request.memberId(), request.result());
+        attendanceRecordService.updateAttendanceRecords(attendanceId, request.memberId(), request.result());
         return ResponseEntity.noContent().build();
     }
 
@@ -95,8 +97,8 @@ public class AttendanceController {
     public ResponseEntity<byte[]> downloadAttendanceRecordsAsExcelBySessions(
             @RequestParam(name = "attendanceIds") List<Long> attendanceIds) {
 
-        byte[] excelFile = attendanceAdminService.createExcelForSessionAttendance(attendanceIds);
-        String finalFileName = attendanceAdminService.getEncodedFileName(attendanceIds);
+        byte[] excelFile = attendanceExcelService.createExcelForSessionAttendance(attendanceIds);
+        String finalFileName = attendanceExcelService.getEncodedFileName(attendanceIds);
 
         HttpHeaders headers = AttendanceExcelHeaderUtil.createExcelDownloadHeaders(finalFileName);
